changed server file to include some socket
