--- public/index.html
@@ -1,24 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-  <meta charset="UTF-8">
-  <meta http-equiv="X-UA-Compatible" content="IE=edge">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>Document</title>
-</head>
-
-
-<body>
-  GAME
-
-  <button type="button" id="join-btn">Join</button>
-
-  <script src="https://cdn.socket.io/4.5.4/socket.io.min.js"
-    integrity="sha384-/KNQL8Nu5gCHLqwqfQjA689Hhoqgi2S84SNUxC3roTe4EhJ9AfLkp8QiQcU8AMzI"
-    crossorigin="anonymous"></script>
-  <script type="module" src="/index.js"></script>
-
-</body>
-
-</html>
